Desktop/Models/Chatbot/requirements.txt
# Core Requirements
torch==2.2.1 --extra-index-url https://download.pytorch.org/whl/cu121  # GPU version
transformers==4.38.2
peft==0.9.0
accelerate==0.27.2
sentencepiece==0.2.0  # Required for Gemma tokenizer

# Deployment
streamlit==1.32.0
python-dotenv==1.0.0
huggingface-hub==0.22.2
bitsandbytes==0.42.0
# Optional (only add if you use these features)
# bitsandbytes==0.42.0  # For 8-bit quantization